IDCT sync with hls-idct (#65)
* DFCxx: added missing Kernel::dfRawBits-method.
* DFCxx: fixed range type for bit extraction on DFVariable and DFVariableImpl.

* Examples, Tests: remade multiple input IDCT example into a singular input; set comparison operations latency to 0 in IDCT config.

* Examples: replaced explicit widths in idct_testbench2.sv with macros.

+++ src/model/dfcxx/lib/dfcxx/vars/var.cpp
@@ -66,9 +66,9 @@ void DFVariableImpl::connect(DFVariableImpl *connectee) {
   meta->graph.addChannel(connectee, this, 0, true);
 }
 
-DFVariableImpl *DFVariableImpl::operator()(uint8_t first, uint8_t second) {
-  uint8_t left = (first > second) ? first : second;
-  uint8_t right = (first > second) ? second : first;
+DFVariableImpl *DFVariableImpl::operator()(uint16_t first, uint16_t second) {
+  uint16_t left = (first > second) ? first : second;
+  uint16_t right = (first > second) ? second : first;
   uint16_t total = getTotalBits();
   if (total <= left) {
     std::stringstream ss;
@@ -227,7 +227,7 @@ void DFVariable::connect(const DFVariable &connectee) {
   impl->connect(connectee.impl);
 }
 
-DFVariable DFVariable::operator()(uint8_t first, uint8_t second) {
+DFVariable DFVariable::operator()(uint16_t first, uint16_t second) {
   return DFVariable(impl->operator()(first, second));
 }
